Support damapper chains
When a damapper .las file with chain flags is given to DaViewer, it now
displays the chains as given (and does infer them).  Absolutely
necessary as such .las files are only sorted with the -a option and so
the chain finding algorithm in DaViewer would fail.

Also found a couple of display bugs affecting chains while I was at it.